pyproject.toml
[tool.poetry]
name = "pytest-kubernetes"
version = "0.1.1"
description = ""
authors = ["Michael Schilonka <michael@blueshoe.io>"]
readme = "README.md"
packages = [{include = "pytest_kubernetes"}]
classifiers = [
    "Framework :: Pytest",
    "Topic :: Software Development :: Testing",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10"
]

[tool.poetry.dependencies]
python = "^3.8.1"
pytest = "^7.2.1"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
coverage = "^7.1.0"
flake8 = "^6.0.0"
flake8-black = "^0.3.6"
mypy = "^1.0.0"
types-pyyaml = "^6.0.12.6"

[tool.poetry.plugins.pytest11]
pytest-kubernetes = "pytest_kubernetes.plugin"

[tool.pytest.ini_options]
markers = [
    "k8s: Kubernetes-based tests",
]

[tool.mypy]
warn_return_any = "True"
warn_unused_configs = "True"
exclude = """
    tests
"""

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
